Barrier Added
İf players hit the barriers, their collected cup which hitted barrier will destroyed. if the hitted cup is the last cup collected just hitted cup going to destroyed. If the hitted cup is in the middle of the collected cups, it will be removed from the list and thrown after the hitted cup.